@@ -1,10 +1,48 @@
 import $ from 'jquery'
-import Chart from 'chart.js'
+import { Chart, LineController, LineElement, PointElement, LinearScale, TimeScale, Title, Tooltip } from 'chart.js'
+import 'chartjs-adapter-moment'
 import humps from 'humps'
 import numeral from 'numeral'
 import moment from 'moment'
 import { formatUsdValue } from '../lib/currency'
 import sassVariables from '../../css/app.scss'
+
+Chart.defaults.font.family = 'Nunito, "Helvetica Neue", Arial, sans-serif,"Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol"'
+Chart.register(LineController, LineElement, PointElement, LinearScale, TimeScale, Title, Tooltip)
+
+const grid = {
+  display: false,
+  drawBorder: false,
+  drawOnChartArea: false
+}
+
+function xAxe (fontColor) {
+  return {
+    grid: grid,
+    type: 'time',
+    time: {
+      unit: 'day',
+      tooltipFormat: 'YYYY-MM-DD',
+      stepSize: 14
+    },
+    ticks: {
+      color: fontColor
+    }
+  }
+}
+
+const padding = {
+  left: 20,
+  right: 20
+}
+
+const legend = {
+  display: false
+}
+
+function formatValue (val) {
+  return `${numeral(val).format('0,0')}`
+}
 
 const config = {
   type: 'line',
@@ -14,81 +52,63 @@ const config = {
   },
   options: {
     layout: {
-      padding: {
-        left: 20,
-        right: 20
-      }
+      padding: padding
     },
-    legend: {
-      display: false
+    interaction: {
+      intersect: false,
+      mode: 'index'
     },
     scales: {
-      xAxes: [{
-        gridLines: {
-          display: false,
-          drawBorder: false
-        },
-        type: 'time',
-        time: {
-          unit: 'day',
-          stepSize: 14
-        },
-        ticks: {
-          fontColor: sassVariables.dashboardBannerChartAxisFontColor
-        }
-      }],
-      yAxes: [{
-        id: 'price',
-        gridLines: {
-          display: false,
-          drawBorder: false
-        },
+      x: xAxe(sassVariables.dashboardBannerChartAxisFontColor),
+      price: {
+        position: 'left',
+        grid: grid,
         ticks: {
           beginAtZero: true,
           callback: (value, _index, _values) => `$${numeral(value).format('0,0.00')}`,
           maxTicksLimit: 4,
-          fontColor: sassVariables.dashboardBannerChartAxisFontColor
+          color: sassVariables.dashboardBannerChartAxisFontColor
         }
-      }, {
-        id: 'marketCap',
-        gridLines: {
-          display: false,
-          drawBorder: false
-        },
+      },
+      marketCap: {
+        position: 'right',
+        grid: grid,
         ticks: {
           callback: (_value, _index, _values) => '',
           maxTicksLimit: 6,
-          drawOnChartArea: false
+          drawOnChartArea: false,
+          color: sassVariables.dashboardBannerChartAxisFontColor
         }
-      }, {
-        id: 'numTransactions',
+      },
+      numTransactions: {
         position: 'right',
-        gridLines: {
-          display: false,
-          drawBorder: false
-        },
+        grid: grid,
         ticks: {
           beginAtZero: true,
-          callback: (value, _index, _values) => `${numeral(value).format('0,0')}`,
+          callback: (value, _index, _values) => formatValue(value),
           maxTicksLimit: 4,
-          fontColor: sassVariables.dashboardBannerChartAxisFontColor
+          color: sassVariables.dashboardBannerChartAxisFontColor
         }
-      }]
+      }
     },
-    tooltips: {
-      mode: 'index',
-      intersect: false,
-      callbacks: {
-        label: ({ datasetIndex, yLabel }, { datasets }) => {
-          const label = datasets[datasetIndex].label
-          if (datasets[datasetIndex].yAxisID === 'price') {
-            return `${label}: ${formatUsdValue(yLabel)}`
-          } else if (datasets[datasetIndex].yAxisID === 'marketCap') {
-            return `${label}: ${formatUsdValue(yLabel)}`
-          } else if (datasets[datasetIndex].yAxisID === 'numTransactions') {
-            return `${label}: ${yLabel}`
-          } else {
-            return yLabel
+    plugins: {
+      legend: legend,
+      tooltip: {
+        mode: 'index',
+        intersect: false,
+        callbacks: {
+          label: (context) => {
+            const { label } = context.dataset
+            const { formattedValue, parsed } = context
+            if (context.dataset.yAxisID === 'price') {
+              return `${label}: ${formatUsdValue(parsed.y)}`
+            } else if (context.dataset.yAxisID === 'marketCap') {
+              return `${label}: ${formatUsdValue(parsed.y)}`
+            } else if (context.dataset.yAxisID === 'numTransactions') {
+              return `${label}: ${formattedValue}`
+            } else {
+              return formattedValue
+            }
           }
         }
       }
@@ -151,11 +171,7 @@ const mcapLineColor = sassVariables.dashboardLineColorMarket
 
 class MarketHistoryChart {
   constructor (el, availableSupply, _marketHistoryData, dataConfig) {
-    const axes = config.options.scales.yAxes.reduce(function (solution, elem) {
-      solution[elem.id] = elem
-      return solution
-    },
-    {})
+    const axes = config.options.scales
 
     let priceActivated = true
     let marketCapActivated = true
@@ -165,6 +181,7 @@ class MarketHistoryChart {
       yAxisID: 'price',
       data: [],
       fill: false,
+      cubicInterpolationMode: 'monotone',
       pointRadius: 0,
       backgroundColor: priceLineColor,
       borderColor: priceLineColor
@@ -181,6 +198,7 @@ class MarketHistoryChart {
       yAxisID: 'marketCap',
       data: [],
       fill: false,
+      cubicInterpolationMode: 'monotone',
       pointRadius: 0,
       backgroundColor: mcapLineColor,
       borderColor: mcapLineColor
@@ -196,6 +214,7 @@ class MarketHistoryChart {
       label: window.localized['Tx/day'],
       yAxisID: 'numTransactions',
       data: [],
+      cubicInterpolationMode: 'monotone',
       fill: false,
       pointRadius: 0,
       backgroundColor: sassVariables.dashboardLineColorTransactions,
@@ -263,16 +282,15 @@ export function createMarketHistoryChart (el) {
             break
           }
           case 'transaction': {
-            const transactionHistory = JSON.parse(data.history_data)
+            const txsHistoryData = JSON.parse(data.history_data)
 
             $(el).show()
-            chart.updateTransactionHistory(transactionHistory)
+            chart.updateTransactionHistory(txsHistoryData)
             break
           }
         }
       })
       .fail(() => {
-        $(el).hide()
         $chartError.show()
       })
   })
