@@ -24,7 +24,8 @@ defmodule BlockScoutWeb.AddressChannel do
     "internal_transaction",
     "transaction",
     "verification_result",
-    "token_transfer"
+    "token_transfer",
+    "pending_transaction"
   ])
 
   {:ok, burn_address_hash} = Chain.string_to_address_hash("0x0000000000000000000000000000000000000000")
@@ -138,6 +139,8 @@ defmodule BlockScoutWeb.AddressChannel do
 
     {:noreply, socket}
   end
+
+  def handle_out("pending_transaction", data, socket), do: handle_transaction(data, socket, "transaction")
 
   def handle_transaction(%{address: address, transaction: transaction}, socket, event) do
     Gettext.put_locale(BlockScoutWeb.Gettext, socket.assigns.locale)
