@@ -11,13 +11,14 @@ defmodule BlockScoutWeb.CSPHeader do
   def call(conn, _opts) do
     Controller.put_secure_browser_headers(conn, %{
       "content-security-policy" => "\
-        connect-src 'self' #{websocket_endpoints(conn)}; \
+        connect-src 'self' #{websocket_endpoints(conn)} https://request-global.czilladx.com/;\
         default-src 'self';\
-        script-src 'self' 'unsafe-inline' 'unsafe-eval';\
+        script-src 'self' 'unsafe-inline' 'unsafe-eval' https://coinzillatag.com;\
         style-src 'self' 'unsafe-inline' 'unsafe-eval' https://fonts.googleapis.com;\
         img-src 'self' * data:;\
         media-src 'self' * data:;\
         font-src 'self' 'unsafe-inline' 'unsafe-eval' https://fonts.gstatic.com data:;\
+        frame-src 'self' 'unsafe-inline' 'unsafe-eval' https://request-global.czilladx.com/;\
       "
     })
   end
