@@ -3,7 +3,7 @@ defmodule BlockScoutWeb.AddressContractView do
 
   alias ABI.{FunctionSelector, TypeDecoder}
   alias Explorer.Chain
-  alias Explorer.Chain.{Address, Data, InternalTransaction, SmartContract}
+  alias Explorer.Chain.{Address, Data, InternalTransaction, SmartContract, Transaction}
 
   def render("scripts.html", %{conn: conn}) do
     render_scripts(conn, "address_contract/code_highlighting.js")
@@ -119,10 +119,19 @@ defmodule BlockScoutWeb.AddressContractView do
     {:ok, contract_code}
   end
 
-  def sourcify_repo_url(address_hash) do
+  def creation_code(%Address{contracts_creation_internal_transaction: %InternalTransaction{}} = address) do
+    address.contracts_creation_internal_transaction.input
+  end
+
+  def creation_code(%Address{contracts_creation_transaction: %Transaction{}} = address) do
+    address.contracts_creation_transaction.input
+  end
+
+  def sourcify_repo_url(address_hash, partial_match) do
     checksummed_hash = Address.checksum(address_hash)
     chain_id = Application.get_env(:explorer, Explorer.ThirdPartyIntegrations.Sourcify)[:chain_id]
     repo_url = Application.get_env(:explorer, Explorer.ThirdPartyIntegrations.Sourcify)[:repo_url]
-    repo_url <> chain_id <> "/" <> checksummed_hash <> "/"
+    match = if partial_match, do: "/partial_match/", else: "/full_match/"
+    repo_url <> match <> chain_id <> "/" <> checksummed_hash <> "/"
   end
 end
