@@ -25,27 +25,13 @@ defmodule Explorer.Chain.Transaction.History.Historian do
       earliest = datetime(day_to_fetch, ~T[00:00:00])
       latest = datetime(day_to_fetch, ~T[23:59:59])
 
-      min_block_query =
+      min_max_block_query =
         from(block in Block,
           where: block.timestamp >= ^earliest and block.timestamp <= ^latest,
-          group_by: block.number,
-          order_by: [asc: min(block.number)],
-          limit: 1,
-          select: min(block.number)
+          select: {min(block.number), max(block.number)}
         )
 
-      min_block = Repo.one(min_block_query, timeout: :infinity)
-
-      max_block_query =
-        from(block in Block,
-          where: block.timestamp >= ^earliest and block.timestamp <= ^latest,
-          group_by: block.number,
-          order_by: [desc: max(block.number)],
-          limit: 1,
-          select: max(block.number)
-        )
-
-      max_block = Repo.one(max_block_query, timeout: :infinity)
+      {min_block, max_block} = Repo.one(min_max_block_query, timeout: :infinity)
 
       if min_block && max_block do
         all_transactions_query =
@@ -59,16 +45,30 @@ defmodule Explorer.Chain.Transaction.History.Historian do
             join: block in Block,
             on: transaction.block_hash == block.hash,
             where: block.consensus == true,
-            select: transaction.hash
+            select: transaction
           )
 
         num_transactions = Repo.aggregate(query, :count, :hash, timeout: :infinity)
         gas_used = Repo.aggregate(query, :sum, :gas_used, timeout: :infinity)
 
-        records = [%{date: day_to_fetch, number_of_transactions: num_transactions, gas_used: gas_used} | records]
+        total_fee_query =
+          from(transaction in subquery(all_transactions_query),
+            join: block in Block,
+            on: transaction.block_hash == block.hash,
+            where: block.consensus == true,
+            select: fragment("SUM(? * ?)", transaction.gas_price, transaction.gas_used)
+          )
+
+        total_fee = Repo.one(total_fee_query, timeout: :infinity)
+
+        records = [
+          %{date: day_to_fetch, number_of_transactions: num_transactions, gas_used: gas_used, total_fee: total_fee}
+          | records
+        ]
+
         compile_records(num_days - 1, records)
       else
-        records = [%{date: day_to_fetch, number_of_transactions: 0} | records]
+        records = [%{date: day_to_fetch, number_of_transactions: 0, gas_used: 0, total_fee: 0} | records]
         compile_records(num_days - 1, records)
       end
     end
