@@ -1174,14 +1174,32 @@ defmodule Explorer.EtherscanTest do
         |> insert()
         |> with_block(first_block)
 
-      insert(:token_transfer, from_address: address, transaction: transaction2)
-      insert(:token_transfer, from_address: address, transaction: transaction1)
-      insert(:token_transfer, from_address: address, transaction: transaction3)
+      insert(:token_transfer,
+        from_address: address,
+        transaction: transaction2,
+        block: transaction2.block,
+        block_number: transaction2.block_number
+      )
+
+      insert(:token_transfer,
+        from_address: address,
+        transaction: transaction1,
+        block: transaction1.block,
+        block_number: transaction1.block_number
+      )
+
+      insert(:token_transfer,
+        from_address: address,
+        transaction: transaction3,
+        block: transaction3.block,
+        block_number: transaction3.block_number
+      )
 
       found_token_transfers = Etherscan.list_token_transfers(address.hash, nil)
 
       block_numbers_order = Enum.map(found_token_transfers, & &1.block_number)
 
+      assert Enum.count(block_numbers_order) == 3
       assert block_numbers_order == Enum.sort(block_numbers_order, &(&1 >= &2))
     end
 
@@ -1206,9 +1224,26 @@ defmodule Explorer.EtherscanTest do
         |> insert()
         |> with_block(first_block)
 
-      insert(:token_transfer, from_address: address, transaction: transaction2)
-      insert(:token_transfer, from_address: address, transaction: transaction1)
-      insert(:token_transfer, from_address: address, transaction: transaction3)
+      insert(:token_transfer,
+        from_address: address,
+        transaction: transaction2,
+        block: transaction2.block,
+        block_number: transaction2.block_number
+      )
+
+      insert(:token_transfer,
+        from_address: address,
+        transaction: transaction1,
+        block: transaction1.block,
+        block_number: transaction1.block_number
+      )
+
+      insert(:token_transfer,
+        from_address: address,
+        transaction: transaction3,
+        block: transaction3.block,
+        block_number: transaction3.block_number
+      )
 
       options = %{order_by_direction: :desc}
 
@@ -1216,6 +1251,7 @@ defmodule Explorer.EtherscanTest do
 
       block_numbers_order = Enum.map(found_token_transfers, & &1.block_number)
 
+      assert Enum.count(block_numbers_order) == 3
       assert block_numbers_order == Enum.sort(block_numbers_order, &(&1 >= &2))
     end
 
@@ -1550,46 +1586,13 @@ defmodule Explorer.EtherscanTest do
 
   describe "get_token_balance/2" do
     test "with a single matching token_balance record" do
-      token_balance =
-        %{token_contract_address_hash: contract_address_hash, address_hash: address_hash} = insert(:token_balance)
+      address_current_token_balance =
+        %{token_contract_address_hash: contract_address_hash, address_hash: address_hash} =
+        insert(:address_current_token_balance)
 
       found_token_balance = Etherscan.get_token_balance(contract_address_hash, address_hash)
 
-      assert found_token_balance.id == token_balance.id
-    end
-
-    test "returns token balance in latest block" do
-      token = insert(:token)
-
-      contract_address_hash = token.contract_address_hash
-
-      address = insert(:address)
-
-      token_details1 = %{
-        token_contract_address_hash: contract_address_hash,
-        address: address,
-        block_number: 5
-      }
-
-      token_details2 = %{
-        token_contract_address_hash: contract_address_hash,
-        address: address,
-        block_number: 15
-      }
-
-      token_details3 = %{
-        token_contract_address_hash: contract_address_hash,
-        address: address,
-        block_number: 10
-      }
-
-      _token_balance1 = insert(:token_balance, token_details1)
-      token_balance2 = insert(:token_balance, token_details2)
-      _token_balance3 = insert(:token_balance, token_details3)
-
-      found_token_balance = Etherscan.get_token_balance(contract_address_hash, address.hash)
-
-      assert found_token_balance.id == token_balance2.id
+      assert found_token_balance.id == address_current_token_balance.id
     end
   end
 
@@ -1598,59 +1601,11 @@ defmodule Explorer.EtherscanTest do
       address = insert(:address)
 
       token_balance =
-        :token_balance
+        :address_current_token_balance
         |> insert(address: address)
         |> Repo.preload(:token)
 
-      insert(:token_balance, address: build(:address))
-
-      token_list = Etherscan.list_tokens(address.hash)
-
-      expected_tokens = [
-        %{
-          balance: token_balance.value,
-          contract_address_hash: token_balance.token_contract_address_hash,
-          name: token_balance.token.name,
-          decimals: token_balance.token.decimals,
-          symbol: token_balance.token.symbol,
-          type: token_balance.token.type
-        }
-      ]
-
-      assert token_list == expected_tokens
-    end
-
-    test "returns the latest known balance per token" do
-      # The latest balance is the one with the latest block number
-      address = insert(:address)
-      token = insert(:token)
-
-      token_balance_details1 = %{
-        address: address,
-        token_contract_address_hash: token.contract_address.hash,
-        block_number: 1
-      }
-
-      token_balance_details2 = %{
-        address: address,
-        token_contract_address_hash: token.contract_address.hash,
-        block_number: 2
-      }
-
-      token_balance_details3 = %{
-        address: address,
-        token_contract_address_hash: token.contract_address.hash,
-        block_number: 3
-      }
-
-      insert(:token_balance, token_balance_details1)
-
-      token_balance =
-        :token_balance
-        |> insert(token_balance_details3)
-        |> Repo.preload(:token)
-
-      insert(:token_balance, token_balance_details2)
+      insert(:address_current_token_balance, address: build(:address))
 
       token_list = Etherscan.list_tokens(address.hash)
 
