@@ -340,8 +340,16 @@ defmodule Indexer.Block.Catchup.Fetcher do
     string_value = Application.get_env(:indexer, :first_block)
 
     case Integer.parse(string_value) do
-      {integer, ""} -> integer
-      _ -> 0
+      {integer, ""} ->
+        integer
+
+      _ ->
+        min_missing_block_number =
+          "min_missing_block_number"
+          |> Chain.get_last_fetched_counter()
+          |> Decimal.to_integer()
+
+        min_missing_block_number
     end
   end
 
