@@ -34,7 +34,7 @@ defmodule Indexer.Fetcher.TokenBalanceOnDemand do
        when not is_nil(address_hash) do
     stale_current_token_balances =
       current_token_balances
-      |> Enum.filter(fn current_token_balance -> current_token_balance.block_number < stale_balance_window end)
+      |> Enum.filter(fn {current_token_balance, _} -> current_token_balance.block_number < stale_balance_window end)
 
     if Enum.count(stale_current_token_balances) > 0 do
       fetch_and_update(latest_block_number, address_hash, stale_current_token_balances)
@@ -48,7 +48,7 @@ defmodule Indexer.Fetcher.TokenBalanceOnDemand do
   defp fetch_and_update(block_number, address_hash, stale_current_token_balances) do
     current_token_balances_update_params =
       stale_current_token_balances
-      |> Enum.map(fn stale_current_token_balance ->
+      |> Enum.map(fn {stale_current_token_balance, _} ->
         stale_current_token_balances_to_fetch = [
           %{
             token_contract_address_hash:
@@ -65,6 +65,7 @@ defmodule Indexer.Fetcher.TokenBalanceOnDemand do
           %{}
           |> Map.put(:address_hash, stale_current_token_balance.address_hash)
           |> Map.put(:token_contract_address_hash, stale_current_token_balance.token_contract_address_hash)
+          |> Map.put(:token_type, stale_current_token_balance.token_type)
           |> Map.put(:block_number, block_number)
           |> Map.put(:value, Decimal.new(updated_balance))
           |> Map.put(:value_fetched_at, DateTime.utc_now())
