@@ -5,6 +5,7 @@ defmodule BlockScout.Mixfile do
 
   def project do
     [
+      app: :block_scout,
       aliases: aliases(Mix.env()),
       version: "2.0",
       apps_path: "apps",
@@ -70,7 +71,7 @@ defmodule BlockScout.Mixfile do
       {:absinthe_plug, git: "https://github.com/blockscout/absinthe_plug.git", tag: "1.5.3", override: true},
       {:tesla, "~> 1.3.3"},
       # Documentation
-      {:ex_doc, "~> 0.19.0", only: [:dev]},
+      {:ex_doc, "~> 0.24.2", only: :dev, runtime: false},
       {:number, "~> 1.0.3"}
     ]
   end
